Move `tapConsole(...)` to `:testutil` module
