+++ logging/src/jvmTest/java/io/spine/logging/JavaWithLoggingTest.java
@@ -31,7 +31,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static com.google.common.truth.Truth.assertThat;
-import static io.spine.testing.logging.TapConsoleKt.tapConsole;
+import static io.spine.logging.testutil.TapConsoleKt.tapConsole;
 
 @DisplayName("In Java, `WithLogging` should")
 class JavaWithLoggingTest {
