+++ logging/src/jvmTest/kotlin/io/spine/logging/LogSiteLookupTest.kt
@@ -5,7 +5,7 @@ import io.kotest.matchers.string.shouldContain
 import io.kotest.matchers.string.shouldNotContain
 import io.spine.logging.LogSiteLookup.logSite
 import io.spine.logging.given.MyLoggingHelper
-import io.spine.testing.logging.tapConsole
+import io.spine.logging.testutil.tapConsole
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
